requirements.txt
yfinance
datetime
seaborn
pandas
numpy
matplotlib
arch
discord.py==1.7.3
dydx-v3-python
python-dotenv